@@ -4,60 +4,60 @@
     "private": true,
     "homepage": "/orkg",
     "dependencies": {
-        "@fortawesome/fontawesome-svg-core": "^1.2.15",
-        "@fortawesome/free-regular-svg-icons": "^5.8.1",
-        "@fortawesome/free-solid-svg-icons": "^5.7.2",
+        "@fortawesome/fontawesome-svg-core": "^1.2.24",
+        "@fortawesome/free-regular-svg-icons": "^5.11.1",
+        "@fortawesome/free-solid-svg-icons": "^5.11.1",
         "@fortawesome/react-fontawesome": "^0.1.4",
         "array-move": "^2.1.0",
         "babel-preset-es2018": "^1.0.0",
         "body-scroll-lock": "^2.6.4",
-        "bootstrap": "^4.1.3",
+        "bootstrap": "^4.3.1",
         "capitalize": "^2.0.0",
-        "citation-js": "^0.4.8",
+        "citation-js": "^0.4.10",
         "classnames": "^2.2.6",
-        "connected-react-router": "^6.3.2",
-        "detect-browser": "^4.6.0",
+        "connected-react-router": "^6.5.2",
+        "detect-browser": "^4.7.0",
         "dot-prop-immutable": "^1.5.0",
         "flow-bin": "^0.95.1",
-        "html2canvas": "^1.0.0-alpha.12",
+        "html2canvas": "1.0.0-alpha.12",
         "jspdf": "^1.5.3",
         "make-latex": "^0.1.0",
         "moment": "^2.24.0",
-        "node-sass": "^4.11.0",
+        "node-sass": "^4.12.0",
         "olap-cube": "^1.0.0",
-        "query-string": "^6.7.0",
+        "query-string": "^6.8.3",
         "randomcolor": "^0.5.4",
         "rangy": "^1.3.0",
-        "react": "^16.8.6",
+        "react": "^16.9.0",
         "react-app-polyfill": "^1.0.2",
-        "react-content-loader": "^4.2.1",
-        "react-contenteditable": "^3.2.6",
-        "react-cookie": "^4.0.0",
+        "react-content-loader": "^4.2.2",
+        "react-contenteditable": "^3.3.2",
+        "react-cookie": "^4.0.1",
         "react-copy-to-clipboard": "^5.0.1",
         "react-csv": "^1.1.1",
-        "react-dom": "^16.6.0",
+        "react-dom": "^16.9.0",
         "react-graph-vis": "github:crubier/react-graph-vis#pull/74/head",
-        "react-hot-loader": "^4.8.2",
+        "react-hot-loader": "^4.12.12",
         "react-input-autosize": "^2.2.1",
-        "react-notifications": "^1.4.3",
-        "react-range": "^1.2.0",
+        "react-range": "^1.3.0",
         "react-redux": "^6.0.1",
         "react-router-config": "^5.0.1",
-        "react-router-dom": "^4.3.1",
-        "react-scripts": "^2.1.8",
+        "react-router-dom": "^5.0.1",
+        "react-scripts": "3.1.2",
         "react-select": "^3.0.4",
-        "react-sortable-hoc": "^1.9.1",
-        "react-split-pane": "^0.1.84",
-        "react-table": "^6.10.0",
+        "react-sortable-hoc": "^1.10.1",
+        "react-split-pane": "^0.1.87",
+        "react-table": "^6.10.3",
         "react-textarea-autosize": "^7.1.0",
-        "react-transition-group": "^1.2.1",
+        "react-toastify": "^5.3.2",
+        "react-transition-group": "^2.9.0",
         "reactour": "^1.15.0",
-        "reactstrap": "^7.1.0",
-        "reactstrap-confirm": "^1.0.4",
-        "redux": "^4.0.1",
+        "reactstrap": "^8.0.1",
+        "reactstrap-confirm": "^1.1.0",
+        "redux": "^4.0.4",
         "redux-thunk": "^2.3.0",
         "strict-mode": "^1.1.3",
-        "styled-components": "^4.1.3",
+        "styled-components": "^4.3.2",
         "validator": "^10.11.0"
     },
     "scripts": {
