@@ -1,16 +1,21 @@
-import React, {Component} from 'react';
+import React, { Component } from 'react';
 import ShortRecord from '../components/statements/ShortRecord';
-import {submitGetRequest, url} from '../network';
+import { submitGetRequest, url } from '../network';
 import { Container } from 'reactstrap';
 
 export default class Predicates extends Component {
-    state = {
-        allPredicates: null,
-        results: null,
-        error: null,
-    };
 
-    async componentWillMount() {
+    constructor(props) {
+        super(props);
+
+        this.state = {
+            allPredicates: null,
+            results: null,
+            error: null,
+        };
+    }
+
+    async componentDidMount() {
         await this.findAllPredicates();
     }
 
@@ -42,7 +47,7 @@ export default class Predicates extends Component {
                 predicate => <ShortRecord key={predicate.id} header={predicate.id}
                     href={`${process.env.PUBLIC_URL}/predicate/${encodeURIComponent(predicate.id)}`}>
                     {predicate.label}
-                    </ShortRecord>
+                </ShortRecord>
             );
 
             return <Container className="box pt-4 pb-4 pl-5 pr-5 mt-5 clearfix">
