@@ -13,14 +13,14 @@ import {
 } from '../../../actions/addPaper';
 import Confirm from 'reactstrap-confirm';
 import Contribution from './Contribution';
-import { CSSTransitionGroup } from 'react-transition-group'
+import { TransitionGroup, CSSTransition } from 'react-transition-group';
 import styled, { withTheme } from 'styled-components';
 import { withCookies } from 'react-cookie';
 import PropTypes from 'prop-types';
 
 
 
-const AnimationContainer = styled.div`
+const AnimationContainer = styled(CSSTransition)`
     transition: 0.3s background-color,  0.3s border-color;
 
     &.fadeIn-enter {
@@ -193,26 +193,22 @@ class Contributions extends Component {
                             </StyledContributionsList>
                         </Col>
 
-                        <CSSTransitionGroup
-                            transitionName="fadeIn"
-                            transitionEnterTimeout={500}
-                            transitionLeave={false}
-                            component="div"
-                            className="col-9"
-                        >
+                        <TransitionGroup className="col-9" exit={false}>
                             <AnimationContainer
+                                classNames="fadeIn"
+                                timeout={{ enter: 500, exit: 0 }}
                                 key={selectedResourceId}
                             >
                                 <Contribution id={selectedResourceId} />
                             </AnimationContainer>
-                        </CSSTransitionGroup>
+                        </TransitionGroup>
                     </Row>
                 </Container>
 
                 <hr className="mt-5 mb-3" />
                 <Button color="primary" className="float-right mb-4" onClick={this.handleNextClick}>Finish</Button>
                 <Button color="light" className="float-right mb-4 mr-2" onClick={this.props.previousStep}>Previous step</Button>
-            </div>
+            </div >
         );
     }
 }
